refactor: remove process-ledger-queue.ts
